backend-pipeline/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
celery>=5.3.0
redis>=4.5.0
flower>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.30.0
django-celery-beat>=2.5.0
python-dotenv>=1.0.0
feedparser>=6.0.0 